+++ src/main/kotlin/com/github/sleepypanda/feesh/features/sounds/MuteJadeDragonSound.kt
@@ -7,11 +7,11 @@ import net.minecraft.resources.ResourceLocation
 object MuteJadeDragonSound {
     @JvmStatic
     fun shouldCancel(soundId: ResourceLocation?): Boolean {
-        if (soundId == null || soundId.namespace != "minecraft" || !soundId.path.startsWith("entity.ender_dragon.")) return false
         if (!WorldRendering.muteJadeDragon) return false
         if (!WorldUtils.isInSkyblock() || WorldUtils.getWorldName() != WorldUtils.GALATEA || 
             (WorldUtils.getZoneName() != WorldUtils.DRAGON_LAIR && WorldUtils.getZoneName() != WorldUtils.MURKWATER_DEPTHS)
         ) return false
+        if (soundId == null || soundId.namespace != "minecraft" || !soundId.path.startsWith("entity.ender_dragon.")) return false
 
         return true
     }